\encoding{utf-8}
\name{sparseMVN-package}
\alias{sparseMVN-package}
\alias{sparseMVN}
\docType{package}
\title{Multivariate normal functions for sparse covariate and precision matrices.}
\description{
 Computes multivariate normal (MVN) densities, and samples from MVN distributions,
for sparse covariance and precision matrices.  This package uses the sparse matrix classes
and algorithms in the Matrix package for computational efficiency.
}
\details{
\tabular{ll}{
Package: \tab sparseMVN\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2013-11-05\cr
License: \tab MPL (>=2.0)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Michael Braun <braunm@smu.edu>
}
\keyword{ package }

