% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{sparseMVN}
\alias{sparseMVN}
\alias{sparseMVN-package}
\title{Multivariate normal functions for sparse covariate and precision matrices.}
\description{
Computes multivariate normal (MVN) densities, and samples
from MVN distributions, for sparse covariance and precision
matrices.  This package uses the sparse matrix classes and
algorithms in the Matrix package for computational
efficiency.
}

