% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMVN.R
\docType{package}
\encoding{UTF-8}
\name{sparseMVN-package}
\alias{sparseMVN-package}
\alias{sparseMVN}
\title{Multivariate Normal Functions for Sparse Covariate and Precision Matrices}
\description{
MVN functions for sparse covariance and precision matrices.
}
\details{
Computes multivariate normal (MVN) densities, and samples
from MVN distributions, when either the covariance or precision
matrix is stored as a sparse Matrix (a dsCMatrix object, as defined
in the Matrix package.  The user can provide the precision matrix
directly, rather than convert it to a covariance via matrix
inversion.
}
\keyword{package}
