% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sparseMVN-package.R
\docType{package}
\encoding{UTF-8}
\name{sparseMVN-package}
\alias{sparseMVN}
\alias{sparseMVN-package}
\title{Multivariate normal functions for sparse covariate and precision matrices.}
\description{
Multivariate normal functions for sparse covariate and
precision matrices.
}
\details{
Computes multivariate normal (MVN) densities, and samples
from MVN distributions, when either the covariance or
precision matrix is stored as a sparse Matrix (a dsCMatrix
object, as defined in the Matrix package.  The user can
provide the precision matrix directly, rather than convert
it to a covariance via matrix inversion.
}
\keyword{package}

