% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMVN.R
\docType{package}
\encoding{UTF-8}
\name{sparseMVN-package}
\alias{sparseMVN-package}
\alias{_PACKAGE}
\alias{sparseMVN}
\title{Multivariate Normal Functions for Sparse Covariate and Precision Matrices}
\description{
MVN functions for sparse covariance and precision matrices.
}
\details{
Computes multivariate normal (MVN) densities, and samples
from MVN distributions, when either the covariance or precision
matrix is stored as a sparse Matrix (a dsCMatrix object, as defined
in the Matrix package.  The user can provide the precision matrix
directly, rather than convert it to a covariance via matrix
inversion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://braunm.github.io/sparseMVN/}
  \item \url{https://github.com/braunm/sparseMVN/}
  \item Report bugs at \url{https://github.com/braunm/sparseMVN/issues/}
}

}
\author{
\strong{Maintainer}: Michael Braun \email{braunm@smu.edu} (\href{https://orcid.org/0000-0003-4774-2119}{ORCID}) [copyright holder]

}
\keyword{package}
